// libs/qb-utils/src/lib/interfaces/WindowSize.ts
export interface WindowSize {
  width: number;
  height: number;
  tiny: boolean;
  small: boolean;
  smallDown: boolean;
  smallUp: boolean;
  medium: boolean;
  mediumDown: boolean;
  mediumUp: boolean;
  large: boolean;
  largeDown: boolean;
  largeUp: boolean;
  xlarge: boolean;
  largeOnly: boolean;
  mediumOnly: boolean;
  smallOnly: boolean;
}